+++ packages/@css-blocks/core/package.json
@@ -37,7 +37,8 @@
     "access": "public"
   },
   "devDependencies": {
-    "@css-blocks/code-style": "^0.20.0-beta.4"
+    "@css-blocks/code-style": "^0.20.0-beta.4",
+    "@css-blocks/test-utils": "^0.20.0-beta.4"
   },
   "dependencies": {
     "@opticss/element-analysis": "^0.4.0",
