+++ packages/@css-blocks/core/test/block-interface-test.ts
@@ -8,8 +8,8 @@ import { setupImporting } from "./util/setupImporting";
 @suite("Block Interfaces")
 export class BlockInterfaceTests extends BEMProcessor {
   @test "can detect missing surface area"() {
-    let { imports, config } = setupImporting();
-    imports.registerSource(
+    let { config, importer } = setupImporting();
+    importer.registerSource(
       "foo/bar/base.css",
       `:scope { color: purple; }
        :scope[state|large] { font-size: 20px; }
@@ -28,20 +28,20 @@ export class BlockInterfaceTests extends BEMProcessor {
       `Missing implementations for: :scope[state|large], .foo[state|small] ` +
         `from foo/bar/base.css`,
       this.process(filename, inputCSS, config).then(() => {
-        imports.assertImported("foo/bar/base.css");
+        importer.assertImported("foo/bar/base.css");
       }));
   }
 
   @test "can import another block"() {
-    let { imports, config } = setupImporting();
-    imports.registerSource(
+    let { config, importer } = setupImporting();
+    importer.registerSource(
       "foo/bar/base.css",
       `:scope { color: purple; }
        :scope[state|large] { font-size: 20px; }
        .foo   { float: left;   }
        .foo[state|small] { font-size: 5px; }`,
     );
-    imports.registerSource(
+    importer.registerSource(
       "foo/bar/other.css",
       `:scope { color: purple; }
       :scope[state|medium] { font-size: 20px; }
@@ -63,8 +63,8 @@ export class BlockInterfaceTests extends BEMProcessor {
       `Missing implementations for: :scope[state|medium], .foo[state|medium] ` +
         `from foo/bar/other.css`,
       this.process(filename, inputCSS, config).then(() => {
-        imports.assertImported("foo/bar/base.css");
-        imports.assertImported("foo/bar/other.css");
+        importer.assertImported("foo/bar/base.css");
+        importer.assertImported("foo/bar/other.css");
       }));
   }
 }
